+++ src/blocks.js
@@ -1731,7 +1731,9 @@ SyntaxElementMorph.prototype.getVarNamesDict = function () {
                 'draggable?' : ['my draggable?'],
                 'rotation style' : ['my rotation style'],
                 'rotation x' : ['my rotation x'],
-                'rotation y' : ['my rotation y']
+                'rotation y' : ['my rotation y'],
+                'scripts' : ['my scripts']
+
             }];
             if (this.world().currentKey === 16) { // shift
                 dict.my[0]['~'] = null; // don't forget we're inside an array...